pcb: Shrink power/ground fill spacing to 0.2mm around FPGA.
The KiCad default here (0.508mm) is probably larger than necessary.  Also
remove most of the redundant traces that were trying to fix the fill not
getting through the gaps between vias.  I could apply this across the entire
board, but that seems like a bad idea for the larger vias.

It might be possible to reduce this further, but it's not worth the risk of an
impossible-to-fix plane-to-via short.